[Chore] #74 - Move SignUpViewModel
SignUp ViewModel 패키지 이동
