+++ app/src/main/java/org/cazait/cazait_android/ui/viewmodel/SignUpViewModel.kt
@@ -1,4 +1,4 @@
-package org.cazait.cazait_android
+package org.cazait.cazait_android.ui.viewmodel
 
 import org.cazait.cazait_android.ui.base.BaseViewModel
